+++ pcs/cli/common/parse_args.py
@@ -14,7 +14,7 @@ PCS_LONG_OPTIONS = [
     "start", "enable", "disabled", "off", "request-timeout=",
     "pacemaker", "corosync",
     "no-default-ops", "defaults", "nodesc",
-    "clone", "master", "name=", "group=", "node=",
+    "master", "name=", "group=", "node=",
     "from=", "to=", "after=", "before=",
     "corosync_conf=",
     "booth-conf=", "booth-key=", "no-watchdog-validation",
@@ -301,27 +301,6 @@ def filter_out_options(arg_list):
         ):
             args_without_options.append(arg)
     return args_without_options
-
-def upgrade_args(arg_list):
-    """
-    Return modified copy of arg_list.
-    This function transform some old syntax to new syntax to keep backward
-    compatibility.
-
-    list arg_list contains command line arguments
-    """
-    upgraded_args = []
-    for arg in arg_list:
-        if arg in ["--cloneopt", "--clone"]:
-            #for every commands - kept as it was previously
-            upgraded_args.append("clone")
-        elif arg.startswith("--cloneopt="):
-            #for every commands - kept as it was previously
-            upgraded_args.append("clone")
-            upgraded_args.append(arg.split('=', 1)[1])
-        else:
-            upgraded_args.append(arg)
-    return upgraded_args
 
 
 class InputModifiers(object):